refactor: share tooltip wrapper between header and cell components
DataTableHeaderCell and DataTableCell mounted identical Fluent Tooltip configurations — same `relationship='description'`, `withArrow`, `mountNode` from the shared provider — with an additional `whiteSpace: 'pre-line'` span on the header for i18n strings that carry `\n`. The cell path didn't have the pre-line style because no cell tooltip currently contains a newline, but the identical-pattern / one-sided-style split is exactly the kind of drift the maintainability review (M-002) flagged: the next time a cell tooltip i18n string gets a `\n`, a developer would need to independently discover and replicate the header's style.

Extract `DataTableTooltip` — a component that owns the mountNode
wiring, the description/arrow config, the pre-line content span,
and the "no wrapper when content is absent" branch. Both call sites
reduce to `<DataTableTooltip content={tooltip}>…</DataTableTooltip>`
with identical behaviour by construction.

Children prop is typed `ReactElement` (not `ReactNode`) because Fluent's Tooltip clones the child to inject ARIA description attributes — fragments or bare text wouldn't round-trip. Both callers pass a single wrapping `div`/`span`, so no behaviour change.
